# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Donut Market Terminal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLoginInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /admin/status:
    get:
      operationId: getAdminStatus
      tags: [admin]
      summary: Admin operational status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminStatus" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/sync:
    post:
      operationId: triggerSync
      tags: [admin]
      summary: Trigger a manual sync
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyncTriggerInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncRunResult" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/watched-players:
    get:
      operationId: getWatchedPlayers
      tags: [admin]
      summary: List watched players
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WatchedPlayer" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: addWatchedPlayer
      tags: [admin]
      summary: Add a watched player
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchedPlayerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchedPlayer" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/watched-players/{username}:
    delete:
      operationId: removeWatchedPlayer
      tags: [admin]
      summary: Remove a watched player
      parameters:
        - { name: username, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/import/preview:
    post:
      operationId: previewImport
      tags: [admin]
      summary: Preview a historical data import
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportReport" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/import/commit:
    post:
      operationId: commitImport
      tags: [admin]
      summary: Commit a historical data import
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportReport" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password: { type: string }
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
    CollectorStatus:
      type: object
      required: [job, status]
      properties:
        job: { type: string }
        status: { type: string }
        lastRunAt: { type: ["string", "null"] }
        lastFinishedAt: { type: ["string", "null"] }
        complete: { type: ["boolean", "null"] }
        recordsInserted: { type: ["integer", "null"] }
        pagesFetched: { type: ["integer", "null"] }
        upstreamRequests: { type: ["integer", "null"] }
        errorSummary: { type: ["string", "null"] }
    TableCount:
      type: object
      required: [table, rows]
      properties:
        table: { type: string }
        rows: { type: integer }
    AdminStatus:
      type: object
      required: [apiKeyConfigured, databaseConfigured, collectors, tableCounts]
      properties:
        apiKeyConfigured: { type: boolean }
        databaseConfigured: { type: boolean }
        requestsPerMinute: { type: integer }
        requestConcurrency: { type: integer }
        orderDataSource: { type: string }
        collectionStartedAt: { type: ["string", "null"] }
        collectors:
          type: array
          items: { $ref: "#/components/schemas/CollectorStatus" }
        tableCounts:
          type: array
          items: { $ref: "#/components/schemas/TableCount" }
    SyncTriggerInput:
      type: object
      required: [job]
      properties:
        job: { type: string, enum: [transactions, listings, leaderboards, rollups, cleanup, all] }
    SyncRunResult:
      type: object
      required: [job, status, recordsInserted, pagesFetched, complete]
      properties:
        job: { type: string }
        status: { type: string }
        recordsInserted: { type: integer }
        pagesFetched: { type: integer }
        upstreamRequests: { type: integer }
        complete: { type: boolean }
        errorSummary: { type: ["string", "null"] }
    WatchedPlayer:
      type: object
      required: [username]
      properties:
        username: { type: string }
        uuid: { type: ["string", "null"] }
        addedAt: { type: ["string", "null"] }
    WatchedPlayerInput:
      type: object
      required: [username]
      properties:
        username: { type: string }
    ImportRow:
      type: object
      required: [itemId, quantity, totalPrice, soldAt]
      properties:
        itemId: { type: string }
        displayName: { type: string }
        quantity: { type: number }
        enchantsJson: { type: string }
        trimJson: { type: string }
        loreJson: { type: string }
        contentsJson: { type: string }
        totalPrice: { type: number }
        sellerName: { type: string }
        sellerUuid: { type: string }
        soldAt: { type: string }
    ImportInput:
      type: object
      required: [rows]
      properties:
        filename: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/ImportRow" }
    ImportRowError:
      type: object
      required: [row, error]
      properties:
        row: { type: integer }
        error: { type: string }
    ImportReport:
      type: object
      required: [rowsRead, accepted, rejected, committed]
      properties:
        rowsRead: { type: integer }
        accepted: { type: integer }
        rejected: { type: integer }
        duplicates: { type: integer }
        committed: { type: boolean }
        errors:
          type: array
          items: { $ref: "#/components/schemas/ImportRowError" }