# lib/api-zod/src/generated/types/importRowError.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Donut Market Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportRowError {
  row: number;
  error: string;
}