# lib/api-zod/src/generated/types/syncRunResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Donut Market Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncRunResult {
  job: string;
  status: string;
  recordsInserted: number;
  pagesFetched: number;
  upstreamRequests?: number;
  complete: boolean;
  /** @nullable */
  errorSummary?: string | null;
}